[Offline mode detection] The app support now detection of offline or online mode, to fetch data from an api or from local database.
[Database]
Model classes and Dao classes have been created, and a converter class for the date format. Also created the RealEstateDatabase class that will be used to access the database.
Objects are inserted from a debug template, and with a debug button in the fragment.

[UI]
RecyclerView layout to display the realty list in the fragment. Support only portrait mode on phone. Landscape and tablet mode will be added later.
The list is automatically updated when an item is inserted into database.

[Debugging]
Added the Stetho debugging tool, mainly to see the database content in real time when the app is running.

[Tests]
The database tests will be added later due to strange failing status (the app can use the database but the tests fail to use it), will be fixed in future versions.