[Persistence] Updated Media reference Dao return type from single item to a list, because a realty can have multiple pictures. Added methods to repository and viewModel to get and insert pictures references and to get a single realty. Tests methods update to match new DAO.
[UI]
Added a fragment to display realty details (in two separate screens for now, tablet support will be added later).
Added navigation actions and arguments to navigation graph to be able to navigate to details fragments with the realty ID in arguments, for Room query.
Added very long strings to see how UI react with huge description (for debug).