[Persistence] Moved AgentID from Realty table to a new table, to allow user to create and update his name when he want to add / update a Realty. DAO Update to reduce code length, using DAO inheritance. DAO Tests update to match new DAO classes.
[Tests]
Assert method removed, because expected failing tests are passing, replaced with assertEquals

[UI]
Removed the online detection in repositories becauce for now there is no online synchronisation, so it is useless. It will be re-added later.
Current user can update his name when he create or update a Realty.

[Other]
Code refactor.