[Api] Removed api Keys from version control. Generated a new api key on Google Cloud Platform. Moved Api key declaration in file apikey.properties, not versioned, added in build.gradle automatic parsing of this file to get api keys available in local project but not available online.
[Persistence]
Added a class that implement ContentProvider class to allow other applications to access to our database. We allow database access only on Realty for now.
The content provider is designed to allow only read-only access to the database.

[BugFix]
Fixed a crash when you try to add a realty for the first time into the database. It was a coroutine problem, in some cases the realty is inserted before the agent, and cause a foreign key constraint fail.
The agent selection is moved in a new settings screen.

[Navigation]
Added a link in menu drawer to to into map and settings screen.

[Documentation]
Added a readme.md file to explain how to compile this project with api keys.

[Other]
Libraries and Gradle update.