[Persistence] Database tables update to match server implementation. DAO update to have ignore or update option with DAO inheritance. Pre fill update, removed RealtyType and Poi prefill.
[Network calls]
Replaced GSON converter with Moshi converter, because it is a best library for Kotlin projects.
Refactor of model classes and created Poi and RealtyTypes models for Ktor server calls.
Added methods to download Poi and RealtyType from server, because for this two classes this is the server that is referent. Poi and RealtyType are common to every instances of application.
Split Google GeoCoding and Ktor retrofit interfaces.

[Debugging tools]
Replaced Facebook Stetho library with Android Debug Database, because facebook tool works only with Google Chrome, and I don't want to have to install Chrome just for this.