[Persistence] Added methods to insert and get PoiRealty information from database.
[UI]
Added checkboxes in AddRealtyFragment.kt, to select some Poi available for a Realty.
Layout updates to fix some elements that wasn't rendered as attempted.
Code optimisation in RealtyViewModel.kt to provide faster result delivery of each screens.

[Bug Fixes]
Now the realty type is correctly displayed in the app.
The edit button in toolbar works in tablet mode.
Icons in toolbar are displayed only on screen that need these icons.